Introducing input
Input now works in that it sends messages to the game and moves the truck. 

However, this change currently freezes the program and does not work correctly. It sends messages too fast and doesn't seem to move left/right appropriately. Because of this, it is commented out for now.

I thought it might be useful to have this on the main file just for any developers that might see this and off help.